[MNG-7324] rename domain to buildinfo and diff to buildsdiff
