+++ maven-core/pom.xml
@@ -215,9 +215,9 @@ under the License.
             <configuration>
               <version>1.0.0</version>
               <models>
+                <model>src/main/mdo/cache-buildinfo.mdo</model>
+                <model>src/main/mdo/cache-buildsdiff.mdo</model>
                 <model>src/main/mdo/cache-config.mdo</model>
-                <model>src/main/mdo/cache-domain.mdo</model>
-                <model>src/main/mdo/cache-diff.mdo</model>
                 <model>src/main/mdo/cache-report.mdo</model>
               </models>
             </configuration>
@@ -233,9 +233,9 @@ under the License.
             <configuration>
               <version>1.0.0</version>
               <models>
+                <model>src/main/mdo/cache-buildinfo.mdo</model>
+                <model>src/main/mdo/cache-buildsdiff.mdo</model>
                 <model>src/main/mdo/cache-config.mdo</model>
-                <model>src/main/mdo/cache-domain.mdo</model>
-                <model>src/main/mdo/cache-diff.mdo</model>
                 <model>src/main/mdo/cache-report.mdo</model>
               </models>
               <outputDirectory>${basedir}/target/generated-sources/modello</outputDirectory>
