+++ maven-core/src/main/java/org/apache/maven/caching/CacheContext.java
@@ -19,7 +19,7 @@ package org.apache.maven.caching;
  * under the License.
  */
 
-import org.apache.maven.caching.xml.domain.ProjectsInputInfoType;
+import org.apache.maven.caching.xml.buildinfo.ProjectsInputInfoType;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.project.MavenProject;
