+++ maven-core/src/main/java/org/apache/maven/caching/CacheControllerImpl.java
@@ -31,14 +31,14 @@ import org.apache.maven.artifact.handler.ArtifactHandler;
 import org.apache.maven.artifact.handler.manager.ArtifactHandlerManager;
 import org.apache.maven.caching.checksum.KeyUtils;
 import org.apache.maven.caching.checksum.MavenProjectInput;
+import org.apache.maven.caching.xml.buildinfo.ArtifactType;
+import org.apache.maven.caching.xml.buildinfo.CompletedExecutionType;
+import org.apache.maven.caching.xml.buildinfo.DigestItemType;
+import org.apache.maven.caching.xml.buildinfo.ProjectsInputInfoType;
+import org.apache.maven.caching.xml.buildinfo.Scm;
+import org.apache.maven.caching.xml.buildsdiff.BuildDiffType;
 import org.apache.maven.caching.xml.config.PropertyNameType;
 import org.apache.maven.caching.xml.config.TrackedPropertyType;
-import org.apache.maven.caching.xml.diff.BuildDiffType;
-import org.apache.maven.caching.xml.domain.ArtifactType;
-import org.apache.maven.caching.xml.domain.CompletedExecutionType;
-import org.apache.maven.caching.xml.domain.DigestItemType;
-import org.apache.maven.caching.xml.domain.ProjectsInputInfoType;
-import org.apache.maven.caching.xml.domain.Scm;
 import org.apache.maven.caching.xml.report.CacheReportType;
 import org.apache.maven.caching.xml.report.ProjectReportType;
 import org.apache.maven.caching.hash.HashAlgorithm;
