+++ maven-core/src/main/java/org/apache/maven/caching/CacheDiff.java
@@ -22,13 +22,13 @@ package org.apache.maven.caching;
 import com.google.common.base.Optional;
 import com.google.common.collect.Sets;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.maven.caching.xml.diff.BuildDiffType;
-import org.apache.maven.caching.xml.diff.MismatchType;
-import org.apache.maven.caching.xml.domain.BuildInfoType;
-import org.apache.maven.caching.xml.domain.CompletedExecutionType;
-import org.apache.maven.caching.xml.domain.DigestItemType;
-import org.apache.maven.caching.xml.domain.ProjectsInputInfoType;
-import org.apache.maven.caching.xml.domain.PropertyValueType;
+import org.apache.maven.caching.xml.buildinfo.BuildInfoType;
+import org.apache.maven.caching.xml.buildinfo.CompletedExecutionType;
+import org.apache.maven.caching.xml.buildinfo.DigestItemType;
+import org.apache.maven.caching.xml.buildinfo.ProjectsInputInfoType;
+import org.apache.maven.caching.xml.buildinfo.PropertyValueType;
+import org.apache.maven.caching.xml.buildsdiff.BuildDiffType;
+import org.apache.maven.caching.xml.buildsdiff.MismatchType;
 import org.apache.maven.caching.xml.CacheConfig;
 
 import java.util.ArrayList;
