+++ maven-core/src/main/java/org/apache/maven/caching/HttpRepositoryImpl.java
@@ -33,8 +33,8 @@ import org.apache.http.entity.InputStreamEntity;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.caching.checksum.MavenProjectInput;
-import org.apache.maven.caching.xml.domain.ArtifactType;
-import org.apache.maven.caching.xml.domain.BuildInfoType;
+import org.apache.maven.caching.xml.buildinfo.ArtifactType;
+import org.apache.maven.caching.xml.buildinfo.BuildInfoType;
 import org.apache.maven.caching.xml.report.CacheReportType;
 import org.apache.maven.caching.xml.report.ProjectReportType;
 import org.apache.maven.caching.xml.BuildInfo;
