+++ maven-core/src/main/java/org/apache/maven/caching/LocalRepositoryImpl.java
@@ -35,9 +35,9 @@ import org.apache.maven.caching.xml.BuildInfo;
 import org.apache.maven.caching.xml.CacheConfig;
 import org.apache.maven.caching.xml.CacheSource;
 import org.apache.maven.caching.xml.XmlService;
-import org.apache.maven.caching.xml.domain.ArtifactType;
-import org.apache.maven.caching.xml.domain.BuildInfoType;
-import org.apache.maven.caching.xml.domain.Scm;
+import org.apache.maven.caching.xml.buildinfo.ArtifactType;
+import org.apache.maven.caching.xml.buildinfo.BuildInfoType;
+import org.apache.maven.caching.xml.buildinfo.Scm;
 import org.apache.maven.caching.xml.report.CacheReportType;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.model.Dependency;
