+++ maven-core/src/main/java/org/apache/maven/caching/ProjectUtils.java
@@ -25,7 +25,7 @@ import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.artifact.handler.ArtifactHandler;
-import org.apache.maven.caching.xml.domain.Scm;
+import org.apache.maven.caching.xml.buildinfo.Scm;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.lifecycle.internal.ProjectIndex;
 import org.apache.maven.lifecycle.internal.builder.BuilderCommon;
