+++ maven-core/src/main/java/org/apache/maven/caching/checksum/MavenProjectInput.java
@@ -38,10 +38,10 @@ import org.apache.maven.caching.hash.HashChecksum;
 import org.apache.maven.caching.xml.BuildInfo;
 import org.apache.maven.caching.xml.CacheConfig;
 import org.apache.maven.caching.xml.DtoUtils;
+import org.apache.maven.caching.xml.buildinfo.DigestItemType;
+import org.apache.maven.caching.xml.buildinfo.ProjectsInputInfoType;
 import org.apache.maven.caching.xml.config.Exclude;
 import org.apache.maven.caching.xml.config.Include;
-import org.apache.maven.caching.xml.domain.DigestItemType;
-import org.apache.maven.caching.xml.domain.ProjectsInputInfoType;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.lifecycle.internal.ProjectIndex;
 import org.apache.maven.model.Build;
