+++ maven-core/src/main/java/org/apache/maven/caching/xml/BuildInfo.java
@@ -24,11 +24,11 @@ import org.apache.maven.artifact.Artifact;
 import org.apache.maven.caching.ProjectUtils;
 import org.apache.maven.caching.checksum.MavenProjectInput;
 import org.apache.maven.caching.hash.HashAlgorithm;
-import org.apache.maven.caching.xml.domain.ArtifactType;
-import org.apache.maven.caching.xml.domain.BuildInfoType;
-import org.apache.maven.caching.xml.domain.CompletedExecutionType;
-import org.apache.maven.caching.xml.domain.DigestItemType;
-import org.apache.maven.caching.xml.domain.ProjectsInputInfoType;
+import org.apache.maven.caching.xml.buildinfo.ArtifactType;
+import org.apache.maven.caching.xml.buildinfo.BuildInfoType;
+import org.apache.maven.caching.xml.buildinfo.CompletedExecutionType;
+import org.apache.maven.caching.xml.buildinfo.DigestItemType;
+import org.apache.maven.caching.xml.buildinfo.ProjectsInputInfoType;
 import org.apache.maven.model.Dependency;
 import org.apache.maven.plugin.MojoExecution;
 import org.codehaus.plexus.logging.Logger;
