+++ maven-core/src/main/mdo/cache-buildsdiff.mdo
@@ -19,15 +19,15 @@ under the License.
 
 <model xmlns="http://codehaus-plexus.github.io/MODELLO/1.4.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://codehaus-plexus.github.io/MODELLO/1.4.0 http://codehaus-plexus.github.io/modello/xsd/modello-1.4.0.xsd"
-  xsd.namespace="org:apache:maven:cache:diff">
-  <id>cache-diff</id>
-  <name>CacheDiff</name>
-  <description><![CDATA[
+  xsd.namespace="org:apache:maven:cache:buildsdiff">
+  <id>cache-buildsdiff</id>
+  <name>CacheBuildsDiff</name>
+  <description><![CDATA[Cache Builds Diff, stored in <code>buildsdiff-${checksum}.xml</code>
   ]]></description>
   <defaults>
     <default>
       <key>package</key>
-      <value>org.apache.maven.caching.xml.diff</value>
+      <value>org.apache.maven.caching.xml.buildsdiff</value>
     </default>
   </defaults>
   <classes>
