+++ maven-core/src/test/java/org/apache/maven/caching/BuildInfoTest.java
@@ -24,12 +24,12 @@ import org.apache.maven.artifact.Artifact;
 import org.apache.maven.artifact.DefaultArtifact;
 import org.apache.maven.artifact.handler.DefaultArtifactHandler;
 import org.apache.maven.caching.hash.HashFactory;
-import org.apache.maven.caching.xml.domain.ArtifactType;
-import org.apache.maven.caching.xml.domain.BuildInfoType;
-import org.apache.maven.caching.xml.domain.CompletedExecutionType;
-import org.apache.maven.caching.xml.domain.DigestItemType;
-import org.apache.maven.caching.xml.domain.ProjectsInputInfoType;
-import org.apache.maven.caching.xml.domain.PropertyValueType;
+import org.apache.maven.caching.xml.buildinfo.ArtifactType;
+import org.apache.maven.caching.xml.buildinfo.BuildInfoType;
+import org.apache.maven.caching.xml.buildinfo.CompletedExecutionType;
+import org.apache.maven.caching.xml.buildinfo.DigestItemType;
+import org.apache.maven.caching.xml.buildinfo.ProjectsInputInfoType;
+import org.apache.maven.caching.xml.buildinfo.PropertyValueType;
 import org.apache.maven.caching.xml.BuildInfo;
 import org.apache.maven.caching.xml.XmlService;
 import org.junit.Test;
